Extract rounded cell view modifier and rounded cell implementations into separate files in the swiftui components section
